Split Home and Summary state test suites #291
