+++ Tests/UnitTests/SummaryRefreshPolicyTests.swift
@@ -1,0 +1,72 @@
+import Foundation
+import CoreLocation
+import SwiftUI
+import Testing
+@testable import SkyAware
+
+@Suite("Outlook Refresh Policy")
+struct OutlookRefreshPolicyTests {
+    private let policy = OutlookRefreshPolicy(minimumSyncInterval: 900)
+
+    @Test("syncs when there is no previous sync")
+    func syncs_withoutPreviousSync() {
+        let now = Date(timeIntervalSince1970: 10_000)
+        #expect(policy.shouldSync(now: now, lastSync: nil, force: false))
+    }
+
+    @Test("skips sync before minimum interval")
+    func skips_beforeMinimumInterval() {
+        let now = Date(timeIntervalSince1970: 10_000)
+        let recent = now.addingTimeInterval(-300)
+        #expect(policy.shouldSync(now: now, lastSync: recent, force: false) == false)
+    }
+
+    @Test("syncs at or beyond minimum interval")
+    func syncs_atOrBeyondMinimumInterval() {
+        let now = Date(timeIntervalSince1970: 10_000)
+        let due = now.addingTimeInterval(-900)
+        #expect(policy.shouldSync(now: now, lastSync: due, force: false))
+    }
+
+    @Test("force refresh bypasses interval guard")
+    func forceBypassesIntervalGuard() {
+        let now = Date(timeIntervalSince1970: 10_000)
+        let recent = now.addingTimeInterval(-5)
+        #expect(policy.shouldSync(now: now, lastSync: recent, force: true))
+    }
+}
+
+
+@Suite("WeatherKit Refresh Policy")
+struct WeatherKitRefreshPolicyTests {
+    private let policy = WeatherKitRefreshPolicy(minimumSyncInterval: 1800)
+
+    @Test("syncs when there is no previous sync")
+    func syncs_withoutPreviousSync() {
+        let now = Date(timeIntervalSince1970: 10_000)
+        #expect(policy.shouldSync(now: now, lastSync: nil, force: false))
+    }
+
+    @Test("skips sync before minimum interval")
+    func skips_beforeMinimumInterval() {
+        let now = Date(timeIntervalSince1970: 10_000)
+        let recent = now.addingTimeInterval(-300)
+        #expect(policy.shouldSync(now: now, lastSync: recent, force: false) == false)
+    }
+
+    @Test("syncs at or beyond minimum interval")
+    func syncs_atOrBeyondMinimumInterval() {
+        let now = Date(timeIntervalSince1970: 10_000)
+        let due = now.addingTimeInterval(-1800)
+        #expect(policy.shouldSync(now: now, lastSync: due, force: false))
+    }
+
+    @Test("force refresh bypasses interval guard")
+    func forceBypassesIntervalGuard() {
+        let now = Date(timeIntervalSince1970: 10_000)
+        let recent = now.addingTimeInterval(-5)
+        #expect(policy.shouldSync(now: now, lastSync: recent, force: true))
+    }
+}
+
+
